pyproject.toml
[tool.poetry]
name = "superdrug-templates"
version = "0.1.0"
description = ""
authors = ["Glenn Goffin <glenn.goffin@solumesl.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
lxml = "^5.2.1"
pydantic = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
